PM / devfreq: dsu-pctrl: update target to be within limits
According to the devfreq_dev_profile target function description, the
device should set its operating frequency at the given freq or
lowest-upper-than-freq value. If freq is higher than any operable
frequency, set the target to maximum.

In the driver's registered target function do not return error for
an invalid target number of portions but set it to an appropriate
value within the limits.

This fixes the use of the performance governor which will not know
the min and max limits at the first update_devfreq triggered by
DEVFREQ_GOV_START, before the devfreq reference is returned by
devfreq_add_device and updated.

Change-Id: I4486c2e9cc5439b05fe9b7cfd8ff5cf4e912ef93
Signed-off-by: Ionela Voinescu <ionela.voinescu@arm.com>