sched/core: add utilization clamping to CPU controller
The CPU cgroup controller allows to assign a specified (maximum)
bandwidth to the tasks of a group. However this bandwidth is defined and
enforced only on a temporal base, without considering the
actual frequency a CPU is run on.
Thus, the amount of computations completed by a task within an allocated
bandwidth can be very different depending on the actual frequency the CPU
is running that task.

With the availability of schedutil, the scheduler has now the possibility
to drive frequency selections based on the actual tasks utilization.
Thus it is now possible to extend the bandwidth controller to specify what
is the maximum or minimum utilization which a task is allowed to generate.
By adding new constraints on minimum and maximum utilization allowed
for tasks in a cpu control group it will be possible to better control the
actual amount of CPU bandwidth consumed by these tasks.

The ultimate goal of these new pair of constraints is to enable:
- tasks boosting: by selecting an higher execution frequency for small
		  tasks which are affecting the user interactive experience
- tasks clamping: by enforcing lower execution frequency (which usually
		  improves energy efficiency) for big tasks which are
		  mainly related to background activities wihtout a direct
		  impact on the user experience.

This patch extends the CPU controller by adding a couple of new
attributes, util_min and util_max, which can be used to enforce
bandwidth boosting and capping. More specifically:

- util_min: defines the minimum CPU utilization which should be considered
            when schedutil selects the frequency for a CPU when a task in
	    this group is RUNNABLE
 	    i.e. the task will run at least at a minimum frequency
                 which corresponds to the min_util utilization

- util_max: defines the maximum CPU utilization which should be considered
	    when schedutil selects the frequency for a CPU when a task in
	    this group is RUNNABLE
	    i.e. the task can run up to a maximum frequency which
	    corresponds to the max_util utilization

These attributes:
a) are tunable at all hierarchy levels, i.e. root group too
b) allow to create subgroups of tasks which are not violating the
   utilization constraints defined by the parent group.

Tasks on a subgroup can only be more boosted and/or more capped, which is
matching with the "limits" schema proposed by the "Resource Distribution
Model (RDM)" suggested by the CGroups v2 documentation:
   Documentation/cgroup-v2.txt

This patch provides the basic support to expose the two new attributes
and to validate their run-time update based on the "limits" schema of
the RDM aforementioned.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Tejun Heo <tj@kernel.org>
Cc: linux-kernel@vger.kernel.org
Cc: linux-pm@vger.kernel.org