sched/core: reference count active tasks's clamp groups
When tasks are enqueue/dequeud from a CPU, the set of clamp groups active
on that CPU can change. Indeed, the clamp value mapped by a clamp group
applies to a CPU only when there is at least one task active in that clamp
group.
Since each clamp group enforce a different utilization clamp value, once
the set of these groups changes it can be required to re-compute what is
the new "aggregated" clamp value to apply for that CPU.

Clamp values are always MAX aggregated for both util_min and util_max. This
is to ensure that no tasks can affect the performances of other
co-scheduled tasks which are either more boosted (i.e.  with higher
util_min clamp) or less capped (i.e. with higher util_max clamp).

This patch introduces the required support to properly reference count
clamp groups at each task enqueue/dequeue time. The MAX aggregation of the
currently active clamp groups is also implemented by reducing at minimum
the number of times we need to scan the complete (unordered) clamp group
array to figure out the new max value.
This "more expensive" operation is limited to only when we dequeue last
task of the clamp group defining the current max clamp, and thus the CPU is
either entering IDLE or going to schedule a less boosted or more clamped
task.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Tejun Heo <tj@kernel.org>
Cc: linux-kernel@vger.kernel.org
Cc: linux-pm@vger.kernel.org