sched: walt: Update cpufreq when a boosted task is enqueued
schedtune updates a CPU's aggregated boost value depending
on the boost values of individual tasks currently on the
runqueue. This happens when a boosted task is enqueued.
Kick cpufreq to potentially raise frequency immediately -
this might be required when the load tracking mechanism
may not initiate a frequency change on its own.

Change-Id: I5f4840d2a2a492c32f4f40416f5f8668cedf0236
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>